requirements.txt
# Dependencies
face_recognition
sklearn
pymongo
python-telegram-bot==12.0.0b1
numpy
python-dateutil
